added pet add function testing
