--- src/main/java/org/nasva/constants/Constants.java
@@ -1,6 +1,0 @@
-package org.nasva.constants;
-
-public class Constants {
-    public static final String BASE_URL = "https://petstore.swagger.io/v2";
-
-}
