+++ src/main/java/org/nasva/models/AddPetDTO.java
@@ -1,11 +1,7 @@
 package org.nasva.models;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import java.util.List;
 import java.util.Objects;
@@ -29,32 +25,4 @@ public class AddPetDTO {
 
     private PetStatus status;
 
-    @Override
-    public int hashCode() {
-        return Objects.hash(
-                name,
-                photoUrls,
-                id,
-                category,
-                tags,
-                status
-        );
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj){
-            return true;
-        }
-        if (!(obj instanceof AddPetDTO)){
-            return false;
-        }
-        AddPetDTO that = (AddPetDTO) obj;
-        return Objects.equals(that.id, this.id) &&
-                Objects.equals(that.name, this.name) &&
-                Objects.equals(that.photoUrls, this.photoUrls) &&
-                that.tags.equals(this.tags) &&
-                that.category.equals(this.category) &&
-                that.status == this.status;
-    }
 }
