+++ src/main/java/org/nasva/models/PetStatus.java
@@ -1,14 +1,8 @@
 package org.nasva.models;
 
-import com.fasterxml.jackson.annotation.JsonValue;
 
-public enum PetStatus{
+public enum PetStatus {
     AVAILABLE,
     PENDING,
-    SOLD;
-
-    @JsonValue
-    public String getValue() {
-        return name().toLowerCase();
-    }
+    SOLD
 }
