+++ src/main/resources/UrlsConfig.properties
@@ -1,0 +1,1 @@
+webapp.urls.baseUrl = https://petstore.swagger.io/v2
