+++ src/test/java/tests/petFeaturesTests/AddPetTests.java
@@ -3,32 +3,37 @@ package tests.petFeaturesTests;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.nasva.generators.pet.PetGenerator;
+import org.nasva.generators.PetGenerator;
 import org.nasva.models.AddPetDTO;
+import org.nasva.models.PetCategoryDTO;
+import org.nasva.models.PetTagDTO;
 import org.nasva.services.PetService;
 
 import java.util.ArrayList;
 import java.util.List;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 public class AddPetTests {
 
     private static final List<AddPetDTO> addedPets = new ArrayList<>();
-    private final int NEGATIVE_ID = -10;
+    private static final int NEGATIVE_ID = -10;
 
     @Test
     public void addPetCorrect(){
         AddPetDTO petRequestDTO = PetGenerator.generateAddPetDto();
         AddPetDTO petResponseDTO = PetService.addPet(petRequestDTO);
-        Assertions.assertEquals(petRequestDTO, petResponseDTO);
+        addedPets.add(petResponseDTO);
+        assertThat(petRequestDTO).usingRecursiveComparison().isEqualTo(petResponseDTO);
     }
 
     @Test
     public void addPetWithNegativeId(){
         AddPetDTO petRequestDTO = PetGenerator.generateAddPetDto(NEGATIVE_ID);
         AddPetDTO petResponseDTO = PetService.addPet(petRequestDTO);
+        addedPets.add(petResponseDTO);
         Assertions.assertTrue(petResponseDTO.getId() > 0);
     }
-
 
     @AfterAll
     public static void deletePets(){
